# Update runtime versions and optimise dotnet implementation