--- a/go/Dockerfile
+++ b/go/Dockerfile
@@ -1,4 +1,4 @@
-FROM golang:1.11.2
+FROM golang:1.14.3
 
 WORKDIR /go/src/app
 COPY . .